Add ImportHygiene sub-class of Lint
